backend/requirements.txt
asgiref==3.10.0
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
inflection==0.5.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
